wrap server in function
